[REFACTOR] decoupling redux-store and visual component
